+ some tests
